Initial Linux/GCC support commit

+++ OTRExporter/OTRExporter/PathExporter.cpp
@@ -9,11 +9,11 @@ void OTRExporter_Path::Save(ZResource* res, const fs::path& outPath, BinaryWrite
 	
 	writer->Write((uint32_t)path->pathways.size());
 
-	for (int k = 0; k < path->pathways.size(); k++)
+	for (size_t k = 0; k < path->pathways.size(); k++)
 	{
 		writer->Write((uint32_t)path->pathways[k].points.size());
 
-		for (int i = 0; i < path->pathways[k].points.size(); i++)
+		for (size_t i = 0; i < path->pathways[k].points.size(); i++)
 		{
 			writer->Write(path->pathways[k].points[i].scalars[0].scalarData.s16);
 			writer->Write(path->pathways[k].points[i].scalars[1].scalarData.s16);
